Make CompilingPass generic over the input only
The return type and error type are supposedly known when defining the
pass. Thus they can be defined as associated types and not
generics. This does not change much to the code of compilation passes,
but it makes pass chaining a lot easier. Instead of using the
silverfish operator with 3 generic parameters (the pass type, the
return type, the error type), it is now used with only one parameter
that is the pass type. The return and error types are determined using
the pass' Residual and Error associated types.